requirements.txt
PySide6==6.8.2.1
PyOpenGL==3.1.9
numpy==2.2.2